requirements.txt
# Core Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0

# Ultralytics YOLO (supports both v8 and v11)
ultralytics>=8.3.0

# SAHI for Slicing Aided Hyper Inference
sahi>=0.11.15

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Experiment Tracking (choose one or both)
wandb>=0.15.0  # Weights & Biases
tensorboard>=2.13.0

# Metrics and Evaluation
scikit-learn>=1.3.0
scipy>=1.11.0

# Progress Bars and Utilities
tqdm>=4.65.0
pyyaml>=6.0

# Jupyter for notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0

# Optional: Albumentations for advanced augmentation
albumentations>=1.3.0

# for Cuda 12.4, uncomment the following line and install via pip
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124